fix: audit & patch semua strict comparison bugs di seluruh alur order
Semua perbandingan user_id/auth()->id() tanpa cast gagal di hosting karena
MySQL PDO mengembalikan nilai kolom sebagai string, bukan integer.

- User/PaymentController: (int) cast pada create() & store(), config key dana.* → services.dana.*
- User/TestimonialController: (int) cast pada create() & store()
- User/TrackingController: (int) cast pada show()
- User/MeasurementController: (int) cast pada destroy()
- Admin/ShipmentController: ganti 'notes' → 'note' (kolom order_statuses)
- Admin/OrderController: ?-> nullsafe pada $order->user->phone
- config/services.php: tambah entry 'dana' (number & name dari env)

+++ app/Http/Controllers/User/MeasurementController.php
@@ -120,7 +120,7 @@ class MeasurementController extends Controller
      */
     public function destroy(Measurement $measurement)
     {
-        if ($measurement->user_id !== auth()->id()) {
+        if ((int) $measurement->user_id !== (int) auth()->id()) {
             abort(403);
         }
